Use for_each to iterate over the project map in the VPC module block to create one VPC for each key/value pair in the map.
This Terraform configuration defines multiple VPCs, assigning each key/value pair in the var.project map to each.key and each.value respectively. When you use for_each with a list or set, each.key is the index of the item in the collection, and each.value is the value of the item.